backend_Intro_Extro/restapi/serializers.py
from rest_framework import serializers
from .models import QuestionAnswer,Answer

class AnswerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Answer
        fields = ('id', 'answer', 'answer_type')

class QuestionAnswerSerializer(serializers.ModelSerializer):
    class Meta:
        model = QuestionAnswer
        fields = ('id', 'question', 'answers')
        depth = 1



