backend_Intro_Extro/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-cors-headers==3.10.1